+++ csp/middleware.py
@@ -1,4 +1,3 @@
-import random
 from functools import partial
 
 from django.conf import settings
@@ -61,15 +60,14 @@ class CSPMiddleware(MiddlewareMixin):
             # Don't overwrite existing headers.
             return response
 
+        response[header] = self.build_policy(request, response)
+
+        return response
+
+    def build_policy(self, request, response):
         config = getattr(response, '_csp_config', None)
         update = getattr(response, '_csp_update', None)
         replace = getattr(response, '_csp_replace', None)
         nonce = getattr(request, '_csp_nonce', None)
-
-        report_percentage = getattr(settings, 'CSP_REPORT_PERCENTAGE', 1)
-        include_report_uri = random.random() < report_percentage
-
-        response[header] = build_policy(config=config, update=update,
-                                        replace=replace, nonce=nonce,
-                                        include_report_uri=include_report_uri)
-        return response
+        return build_policy(config=config, update=update, replace=replace,
+                            nonce=nonce)
