+++ docs/configuration.rst
@@ -64,11 +64,6 @@ These settings affect the policy in the header. The defaults are in
     Set the ``report-uri`` directive. A **string** with a full or
     relative URI.
     Note: This doesn't use default-src as a fall-back.
-``CSP_REPORT_PERCENTAGE``
-    Percentage of requests that should see the ``report-uri`` directive.
-    Use this to throttle the number of CSP violation reports made to your
-    ``CSP_REPORT_URI``. A **float** between 0 and 1 (0 = no reports at all).
-    Ignored if ``CSP_REPORT_URI`` isn't set.
 ``CSP_MANIFEST_SRC``
     Set the ``manifest-src`` directive. A tuple or list. *None*
 ``CSP_WORKER_SRC``
